Updated timeline cards and improved sizing in mobile. Refactored timeline cards components and composition
