+++ Apps/WebClient/src/ClientApp/src/components/plot/lineChart.vue
