+++ Apps/WebClient/src/ClientApp/src/components/report/mspVisits.vue
@@ -183,22 +183,22 @@ export default class MSPVisitsReportComponent extends Vue {
                     <b-col>No records found.</b-col>
                 </b-row>
                 <b-row v-else-if="!isEmpty" class="py-3 mt-4 header">
-                    <b-col class="col">Date</b-col>
-                    <b-col class="col">Provider Name</b-col>
-                    <b-col class="col">Specialty Description</b-col>
-                    <b-col class="col">Clinic Location</b-col>
+                    <b-col>Date</b-col>
+                    <b-col>Provider Name</b-col>
+                    <b-col>Specialty Description</b-col>
+                    <b-col>Clinic Location</b-col>
                 </b-row>
                 <b-row v-for="item in records" :key="item.id" class="item py-1">
-                    <b-col class="col my-auto text-nowrap">
+                    <b-col class="my-auto text-nowrap">
                         {{ formatDate(item.encounterDate) }}
                     </b-col>
-                    <b-col class="col my-auto">
+                    <b-col class="my-auto">
                         {{ item.practitionerName }}
                     </b-col>
-                    <b-col class="col my-auto">
+                    <b-col class="my-auto">
                         {{ item.specialtyDescription }}
                     </b-col>
-                    <b-col class="col my-auto">
+                    <b-col class="my-auto">
                         <p>{{ item.clinic.name }}</p>
                         <p>{{ item.clinic.address }}</p>
                     </b-col>
