--- Apps/WebClient/src/ClientApp/src/components/timeline/entrycard.vue
@@ -1,88 +1,0 @@
-<script lang="ts">
-import Vue from "vue";
-import { Component, Prop } from "vue-property-decorator";
-
-import TimelineEntry, { EntryType } from "@/models/timelineEntry";
-
-import EncounterTimelineComponent from "./encounter.vue";
-import ImmunizationTimelineComponent from "./immunization.vue";
-import LaboratoryTimelineComponent from "./laboratory.vue";
-import MedicationTimelineComponent from "./medication.vue";
-import NoteTimelineComponent from "./note.vue";
-
-@Component({
-    components: {
-        MedicationComponent: MedicationTimelineComponent,
-        ImmunizationComponent: ImmunizationTimelineComponent,
-        LaboratoryComponent: LaboratoryTimelineComponent,
-        EncounterComponent: EncounterTimelineComponent,
-        NoteComponent: NoteTimelineComponent,
-    },
-})
-export default class EntrycardTimelineComponent extends Vue {
-    @Prop() datekey!: string;
-    @Prop() entry!: TimelineEntry;
-    @Prop() index!: number;
-
-    private get EntryType(): typeof EntryType {
-        return EntryType;
-    }
-}
-</script>
-
-<template>
-    <b-row class="entryCard p-0 m-0">
-        <MedicationComponent
-            v-if="entry.type === EntryType.Medication"
-            :key="entry.id"
-            :datekey="datekey"
-            :entry="entry"
-            :index="index"
-        />
-        <ImmunizationComponent
-            v-if="entry.type === EntryType.Immunization"
-            :key="entry.id"
-            :datekey="datekey"
-            :entry="entry"
-            :index="index"
-        />
-        <LaboratoryComponent
-            v-if="entry.type === EntryType.Laboratory"
-            :key="entry.id"
-            :datekey="datekey"
-            :entry="entry"
-            :index="index"
-        />
-        <EncounterComponent
-            v-if="entry.type === EntryType.Encounter"
-            :key="entry.id"
-            :datekey="datekey"
-            :entry="entry"
-            :index="index"
-        />
-        <NoteComponent
-            v-if="entry.type === EntryType.Note"
-            :key="entry.id"
-            :datekey="datekey"
-            :entry="entry"
-            :index="index"
-        />
-    </b-row>
-</template>
-
-<style lang="scss" scoped>
-.entryCard {
-    width: 100%;
-    min-width: 100%;
-    /*padding-top: 10px;
-    padding-bottom: 10px;*/
-    /*border-color: red;
-    border-width: 1px;
-    border-style: solid;*/
-}
-/*@media screen and (max-width: 600px) {
-    .entryCard {
-        padding-left: 30px;
-    }
-}*/
-</style>
