--- Apps/WebClient/src/ClientApp/src/components/timeline/laboratory.vue
@@ -1,360 +1,0 @@
-<script lang="ts">
-import { library } from "@fortawesome/fontawesome-svg-core";
-import { faFlask, IconDefinition } from "@fortawesome/free-solid-svg-icons";
-import { faFileDownload } from "@fortawesome/free-solid-svg-icons";
-import Vue from "vue";
-import { Component, Prop, Ref } from "vue-property-decorator";
-import { Getter } from "vuex-class";
-
-import MessageModalComponent from "@/components/modal/genericMessage.vue";
-import CommentSectionComponent from "@/components/timeline/commentSection.vue";
-import { DateWrapper } from "@/models/dateWrapper";
-import { LaboratoryReport } from "@/models/laboratory";
-import LaboratoryTimelineEntry from "@/models/laboratoryTimelineEntry";
-import User from "@/models/user";
-import { SERVICE_IDENTIFIER } from "@/plugins/inversify";
-import container from "@/plugins/inversify.config";
-import { ILaboratoryService } from "@/services/interfaces";
-library.add(faFileDownload);
-
-@Component({
-    components: {
-        MessageModalComponent,
-        CommentSection: CommentSectionComponent,
-    },
-})
-export default class LaboratoryTimelineComponent extends Vue {
-    @Prop() entry!: LaboratoryTimelineEntry;
-    @Prop() index!: number;
-    @Prop() datekey!: string;
-    @Getter("user", { namespace: "user" }) user!: User;
-
-    @Ref("messageModal")
-    readonly messageModal!: MessageModalComponent;
-
-    private laboratoryService!: ILaboratoryService;
-
-    private hasErrors = false;
-    private detailsVisible = false;
-    private isLoadingDocument = false;
-
-    private mounted() {
-        this.laboratoryService = container.get<ILaboratoryService>(
-            SERVICE_IDENTIFIER.LaboratoryService
-        );
-    }
-
-    private get entryIcon(): IconDefinition {
-        return faFlask;
-    }
-
-    private toggleDetails(): void {
-        this.detailsVisible = !this.detailsVisible;
-        this.hasErrors = false;
-    }
-
-    private formatDate(date: DateWrapper): string {
-        return date.format("LLL dd, yyyy t");
-    }
-
-    private showConfirmationModal(): void {
-        this.messageModal.showModal();
-    }
-
-    private getReport() {
-        this.isLoadingDocument = true;
-        this.laboratoryService
-            .getReportDocument(this.entry.id, this.user.hdid)
-            .then((result) => {
-                const link = document.createElement("a");
-                let dateString = this.entry.displayDate.format(
-                    "YYYY_MM_DD-HH_mm"
-                );
-                let report: LaboratoryReport = result.resourcePayload;
-                link.href = `data:${report.mediaType};${report.encoding},${report.data}`;
-                link.download = `COVID_Result_${dateString}.pdf`;
-                link.click();
-                URL.revokeObjectURL(link.href);
-            })
-            .catch(() => {
-                this.hasErrors = true;
-            })
-            .finally(() => {
-                this.isLoadingDocument = false;
-            });
-    }
-
-    private get reportAvailable(): boolean {
-        return this.entry.reportAvailable;
-    }
-}
-</script>
-
-<template>
-    <b-col class="timelineCard">
-        <b-row class="entryHeading px-2 py-3">
-            <b-col class="leftPane">
-                <div class="icon">
-                    <font-awesome-icon
-                        :icon="entryIcon"
-                        size="lg"
-                    ></font-awesome-icon>
-                </div>
-            </b-col>
-            <b-col class="entryTitle">
-                <b-row class="justify-content-between">
-                    <b-col data-testid="laboratoryTitle">
-                        <strong>{{ entry.summaryTitle }}</strong>
-                    </b-col>
-                </b-row>
-                <b-row class="py-2">
-                    <b-col class="text-muted">
-                        <strong v-show="entry.isStatusFinal">
-                            Result:
-                            <span :class="entry.labResultOutcome"
-                                >{{ entry.labResultOutcome }}
-                            </span></strong
-                        >
-                    </b-col>
-                    <b-col class="rightPane">
-                        <div>
-                            <b-btn
-                                variant="link"
-                                class="detailsButton"
-                                @click="toggleDetails()"
-                            >
-                                <span v-if="detailsVisible">Hide Details</span>
-                                <span v-else>View Details</span>
-                            </b-btn>
-                        </div>
-                    </b-col>
-                </b-row>
-            </b-col>
-        </b-row>
-        <b-row>
-            <b-col class="leftPane d-none d-md-block"></b-col>
-            <b-col>
-                <b-collapse
-                    :id="'entryDetails-' + index + '-' + datekey"
-                    v-model="detailsVisible"
-                >
-                    <b-row class="justify-content-between">
-                        <b-col cols="8">
-                            {{ entry.summaryDescription }}
-                        </b-col>
-                        <b-col v-if="reportAvailable" cols="auto" class="pr-0">
-                            <b-spinner v-if="isLoadingDocument"></b-spinner>
-                            <span v-else>
-                                <strong>Report:</strong>
-                                <b-btn
-                                    variant="link"
-                                    @click="showConfirmationModal()"
-                                >
-                                    <font-awesome-icon
-                                        icon="file-download"
-                                        aria-hidden="true"
-                                        size="1x"
-                                    />
-                                </b-btn>
-                            </span>
-
-                            <div class="detailSection">
-                                <div>
-                                    <strong>Ordering Providers:</strong>
-                                    {{ entry.orderingProviders }}
-                                </div>
-                                <div>
-                                    <strong>Reporting Lab:</strong>
-                                    {{ entry.reportingLab }}
-                                </div>
-                                <div>
-                                    <strong>Location:</strong>
-                                    {{ entry.location }}
-                                </div>
-                            </div>
-
-                            <div class="detailSection">
-                                <strong>Results:</strong>
-                                <div
-                                    v-for="result in entry.resultList"
-                                    :key="result.id"
-                                    class="border p-1"
-                                >
-                                    <div>
-                                        <strong>Test Type:</strong>
-                                        {{ result.testType }}
-                                    </div>
-                                    <div>
-                                        <strong>Out Of Range:</strong>
-                                        {{ result.outOfRange }}
-                                    </div>
-                                    <div>
-                                        <strong>Test Status:</strong>
-                                        {{ result.testStatus }}
-                                    </div>
-                                    <div class="my-2">
-                                        <strong>Result Description:</strong>
-                                        <p
-                                            v-html="result.resultDescription"
-                                        ></p>
-                                    </div>
-                                    <div>
-                                        <strong>Collected Date Time:</strong>
-                                        {{
-                                            formatDate(result.collectedDateTime)
-                                        }}
-                                    </div>
-
-                                    <div>
-                                        <strong>Received Date Time:</strong>
-                                        {{
-                                            formatDate(result.receivedDateTime)
-                                        }}
-                                    </div>
-                                </div>
-                            </div>
-
-                            <div class="detailSection">
-                                <div>
-                                    <strong>What to expect next</strong>
-                                    <p>
-                                        If you receive a
-                                        <b>positive</b> COVID-19 result:
-                                    </p>
-                                    <ul>
-                                        <li>
-                                            You and the people you live with
-                                            need to self-isolate now.
-                                        </li>
-                                        <li>
-                                            Public health will contact you in 2
-                                            to 3 days with further instructions.
-                                        </li>
-                                        <li>
-                                            If you are a health care worker,
-                                            please notify your employer of your
-                                            positive result.
-                                        </li>
-                                        <li>
-                                            Monitor your health and contact a
-                                            health care provider or call 8-1-1
-                                            if you are concerned about your
-                                            symptoms.
-                                        </li>
-                                        <li>
-                                            Go to
-                                            <a
-                                                href="http://www.bccdc.ca/results"
-                                                target="blank_"
-                                                >www.bccdc.ca/results</a
-                                            >
-                                            for more information about your test
-                                            result.
-                                        </li>
-                                    </ul>
-                                </div>
-                            </div>
-                        </b-col>
-                    </b-row>
-                    <CommentSection :parent-entry="entry"></CommentSection>
-                </b-collapse>
-            </b-col>
-        </b-row>
-        <MessageModalComponent
-            ref="messageModal"
-            title="Sensitive Document Download"
-            message="The file that you are downloading contains personal information. If you are on a public computer, please ensure that the file is deleted before you log off."
-            @submit="getReport"
-        />
-    </b-col>
-</template>
-
-<style lang="scss" scoped>
-@import "@/assets/scss/_variables.scss";
-
-$radius: 15px;
-
-$left-pane-width: 80px;
-
-.col {
-    padding: 0px;
-    margin: 0px;
-}
-.row {
-    padding: 0;
-    margin: 0px;
-}
-
-.timelineCard {
-    //border-color: green;
-    border-radius: $radius $radius $radius $radius;
-    border-color: $soft_background;
-    border-style: solid;
-    border-width: 2px;
-}
-
-.entryHeading {
-    background-color: $soft_background;
-}
-
-.entryTitle {
-    color: $primary;
-    text-align: left;
-}
-
-/*.icon {
-    background-color: $primary;
-    color: white;
-    text-align: center;
-    padding: 10px 0;
-    border-radius: $radius 0px 0px 0px;
-}*/
-
-.icon {
-    background-color: $primary;
-    color: white;
-    text-align: center;
-    border-radius: 50%;
-    height: 60px;
-    width: 60px;
-    padding-top: 17px;
-    font-size: 1.2em;
-    //margin: 10px;
-}
-
-.leftPane {
-    width: $left-pane-width;
-    max-width: $left-pane-width;
-}
-$right-pane-width: 90px;
-.rightPane {
-    text-align: center;
-    width: $right-pane-width;
-    max-width: $right-pane-width;
-}
-
-.detailsButton {
-    padding: 0px;
-    text-align: right;
-    //margin-top: 50%;
-}
-
-.detailSection {
-    margin-top: 15px;
-}
-
-.commentButton {
-    border-radius: $radius;
-}
-
-.newComment {
-    border-radius: $radius;
-}
-
-span.Positive {
-    color: red;
-}
-span.Negative {
-    color: green;
-}
-</style>
