+++ Apps/WebClient/src/ClientApp/src/views/timeline.vue
@@ -8,12 +8,11 @@ import { Action, Getter } from "vuex-class";
 import ErrorCardComponent from "@/components/errorCard.vue";
 import LoadingComponent from "@/components/loading.vue";
 import CovidModalComponent from "@/components/modal/covid.vue";
+import NoteEditComponent from "@/components/modal/noteEdit.vue";
 import ProtectiveWordComponent from "@/components/modal/protectiveWord.vue";
 import CalendarTimelineComponent from "@/components/timeline/calendarTimeline.vue";
-import EntryCardTimelineComponent from "@/components/timeline/entrycard.vue";
 import FilterComponent from "@/components/timeline/filters.vue";
 import LinearTimelineComponent from "@/components/timeline/linearTimeline.vue";
-import NoteTimelineComponent from "@/components/timeline/note.vue";
 import { ActionType } from "@/constants/actionType";
 import { ResultType } from "@/constants/resulttype";
 import UserPreferenceType from "@/constants/userPreferenceType";
@@ -54,8 +53,7 @@ Component.registerHooks(["beforeRouteLeave"]);
         LoadingComponent,
         ProtectiveWordComponent,
         CovidModalComponent,
-        EntryCardComponent: EntryCardTimelineComponent,
-        NoteTimelineComponent,
+        NoteEditComponent,
         LinearTimeline: LinearTimelineComponent,
         CalendarTimeline: CalendarTimelineComponent,
         ErrorCard: ErrorCardComponent,
@@ -717,91 +715,93 @@ export default class TimelineView extends Vue {
 <template>
     <div>
         <LoadingComponent v-if="isLoading" :is-custom="true"></LoadingComponent>
-        <b-row class="my-3 fluid p-0 m-0">
-            <b-col id="timeline" class="col-12 col-lg-9 column-wrapper m-0 p-0">
-                <b-alert
-                    :show="hasNewTermsOfService"
-                    dismissible
-                    variant="info"
-                    class="no-print"
-                >
-                    <h4>Updated Terms of Service</h4>
-                    <span>
-                        The Terms of Service have been updated since your last
-                        login. You can review them
-                        <router-link
-                            id="termsOfServiceLink"
-                            variant="primary"
-                            to="/termsOfService"
-                        >
-                            here</router-link
-                        >.
-                    </span>
-                </b-alert>
-                <b-alert
-                    :show="unverifiedEmail || unverifiedSMS"
-                    dismissible
-                    variant="info"
-                    class="no-print"
-                >
-                    <h4>Please complete your profile</h4>
-                    <span>
-                        Your email or cell phone number have not been verified.
-                        To complete your profile and receive notifications from
-                        the Health Gateway, visit the
-                        <router-link
-                            id="profilePageLink"
-                            variant="primary"
-                            to="/profile"
-                            >Profile Page</router-link
-                        >
-                        <span>.</span>
-                    </span>
-                </b-alert>
-                <b-alert
-                    :show="!isPacificTime"
-                    dismissible
-                    variant="info"
-                    class="no-print"
-                >
-                    <h4>Looks like you're in a different timezone.</h4>
-                    <span>
-                        Heads up: your health records are recorded and displayed
-                        in Pacific Time.
-                    </span>
-                </b-alert>
-                <b-alert
-                    :show="immunizationLoadDeferred"
-                    variant="info"
-                    class="no-print"
-                >
-                    <span v-if="!immunizationLoadReady">
-                        <h4 data-testid="immunizationLoading">
-                            Still loading your immunization records
-                        </h4>
-                    </span>
-                    <span v-else data-testid="immunizationReady">
-                        <h4 data-testid="immunizationReadyHeader">
-                            Your immunization records are ready
-                        </h4>
-                        <b-btn
-                            data-testid="immunizationBtnReady"
-                            variant="link"
-                            class="detailsButton px-0"
-                            @click="loadImmunizationEntries()"
-                        >
-                            Load to timeline.
-                        </b-btn></span
+        <b-row class="my-2 fluid">
+            <b-col id="timeline" class="col-12 col-lg-9 column-wrapper">
+                <div class="px-2">
+                    <b-alert
+                        :show="hasNewTermsOfService"
+                        dismissible
+                        variant="info"
+                        class="no-print"
                     >
-                </b-alert>
+                        <h4>Updated Terms of Service</h4>
+                        <span>
+                            The Terms of Service have been updated since your
+                            last login. You can review them
+                            <router-link
+                                id="termsOfServiceLink"
+                                variant="primary"
+                                to="/termsOfService"
+                            >
+                                here</router-link
+                            >.
+                        </span>
+                    </b-alert>
+                    <b-alert
+                        :show="unverifiedEmail || unverifiedSMS"
+                        dismissible
+                        variant="info"
+                        class="no-print"
+                    >
+                        <h4>Please complete your profile</h4>
+                        <span>
+                            Your email or cell phone number have not been
+                            verified. To complete your profile and receive
+                            notifications from the Health Gateway, visit the
+                            <router-link
+                                id="profilePageLink"
+                                variant="primary"
+                                to="/profile"
+                                >Profile Page</router-link
+                            >
+                            <span>.</span>
+                        </span>
+                    </b-alert>
+                    <b-alert
+                        :show="!isPacificTime"
+                        dismissible
+                        variant="info"
+                        class="no-print"
+                    >
+                        <h4>Looks like you're in a different timezone.</h4>
+                        <span>
+                            Heads up: your health records are recorded and
+                            displayed in Pacific Time.
+                        </span>
+                    </b-alert>
+                    <b-alert
+                        :show="immunizationLoadDeferred"
+                        variant="info"
+                        class="no-print"
+                    >
+                        <span v-if="!immunizationLoadReady">
+                            <h4 data-testid="immunizationLoading">
+                                Still loading your immunization records
+                            </h4>
+                        </span>
+                        <span v-else data-testid="immunizationReady">
+                            <h4 data-testid="immunizationReadyHeader">
+                                Your immunization records are ready
+                            </h4>
+                            <b-btn
+                                data-testid="immunizationBtnReady"
+                                variant="link"
+                                class="detailsButton px-0"
+                                @click="loadImmunizationEntries()"
+                            >
+                                Load to timeline.
+                            </b-btn></span
+                        >
+                    </b-alert>
+                </div>
 
-                <div id="pageTitle">
+                <div id="pageTitle" class="px-2">
                     <h1 id="subject">Health Care Timeline</h1>
                     <hr class="mb-0" />
                 </div>
-                <div class="sticky-top sticky-offset">
+                <div class="sticky-top sticky-offset px-2">
                     <b-row class="no-print justify-content-between">
-                        <b-col class="col">
+                        <b-col>
                             <div class="form-group has-filter">
                                 <font-awesome-icon
                                     :icon="searchIcon"
@@ -818,7 +818,7 @@ export default class TimelineView extends Vue {
                                 ></b-form-input>
                             </div>
                         </b-col>
-                        <b-col v-if="!isLoading" class="col-auto pl-0">
+                        <b-col v-if="!isLoading" class="col-auto pl-2">
                             <Filters
                                 :is-list-view="isListView"
                                 :filter.sync="filter"
@@ -837,32 +837,7 @@ export default class TimelineView extends Vue {
                     :is-visible="isListView"
                     :total-entries="getTotalCount()"
                     :filter="filter"
-                >
-                    <b-row
-                        slot="month-list-toggle"
-                        class="view-selector justify-content-end"
-                    >
-                        <b-col cols="auto" class="pr-0">
-                            <b-btn
-                                data-testid="monthViewToggle"
-                                class="month-view-btn btn-outline-primary px-2 m-0"
-                                :class="{ active: false }"
-                                @click.stop="toggleMonthView"
-                            >
-                                Date
-                            </b-btn>
-                        </b-col>
-                        <b-col cols="auto" class="pl-0">
-                            <b-btn
-                                data-testid="listViewToggle"
-                                class="list-view-btn btn-outline-primary px-2 m-0"
-                                :class="{ active: true }"
-                            >
-                                List
-                            </b-btn>
-                        </b-col>
-                    </b-row>
-                </LinearTimeline>
+                />
                 <CalendarTimeline
                     v-show="!isListView && !isLoading"
                     :timeline-entries="timelineEntries"
@@ -930,21 +905,53 @@ export default class TimelineView extends Vue {
             @submit="onProtectiveWordSubmit"
             @cancel="onProtectiveWordCancel"
         />
+        <NoteEditComponent
+            ref="noteEditModal"
+            :is-loading="isLoading"
+            @submit="onCovidSubmit"
+            @cancel="onCovidCancel"
+        />
     </div>
 </template>
 
 <style lang="scss" scoped>
 @import "@/assets/scss/_variables.scss";
+.row {
+    margin: 0px;
+    padding: 0px;
+}
+
+.col {
+    margin: 0px;
+    padding: 0px;
+}
+
 .column-wrapper {
     border: 1px;
 }
 
 #pageTitle {
     color: $primary;
+
+    hr {
+        border-top: 2px solid $primary;
+    }
+
+    h1 {
+        @media (max-width: 575px) {
+            font-size: 2em !important;
+        }
+    }
 }
 
-#pageTitle hr {
-    border-top: 2px solid $primary;
+.form-group {
+    margin-bottom: 0px;
+}
+
+.sticky-offset {
+    padding-top: 1rem;
+    background-color: white;
+    z-index: 2;
 }
 
 .has-filter .form-control {
@@ -998,10 +1005,5 @@ export default class TimelineView extends Vue {
 
 .sticky-top {
     z-index: 49 !important;
-}
-.sticky-offset {
-    padding-top: 1rem;
-    background-color: white;
-    z-index: 2;
 }
 </style>
